+++ Sources/LiveKit/Track/Capturers/CameraCapturer.swift
@@ -22,57 +22,6 @@ import ReplayKit
 
 @_implementationOnly import LiveKitWebRTC
 
-// Internal-only for now
-class DeviceManager: Loggable {
-    // MARK: - Public
-
-    public static let shared = DeviceManager()
-
-    public var devices: [AVCaptureDevice] { _state.devices }
-
-    struct State {
-        var devices: [AVCaptureDevice] = []
-    }
-
-    private var _state = StateSync(State())
-
-    public lazy var session: AVCaptureDevice.DiscoverySession = {
-        let deviceTypes: [AVCaptureDevice.DeviceType]
-        #if os(iOS)
-        deviceTypes = [
-            .builtInDualCamera,
-            .builtInDualWideCamera,
-            .builtInTripleCamera,
-            .builtInWideAngleCamera,
-            .builtInTelephotoCamera,
-            .builtInUltraWideCamera,
-        ]
-        #else
-        deviceTypes = [
-            .builtInWideAngleCamera,
-        ]
-        #endif
-
-        return AVCaptureDevice.DiscoverySession(deviceTypes: deviceTypes,
-                                                mediaType: .video,
-                                                position: .unspecified)
-    }()
-
-    var _observation: NSKeyValueObservation?
-
-    init() {
-        log()
-
-        _observation = session.observe(\.devices, options: [.initial, .new]) { [weak self] _, value in
-            guard let self else { return }
-            self.log("Devices: \(String(describing: value.newValue))")
-            self._state.mutate {
-                $0.devices = value.newValue ?? []
-            }
-        }
-    }
-}
-
 public class CameraCapturer: VideoCapturer {
     @objc
     public static func captureDevices() -> [AVCaptureDevice] {
